Move blocks range functionality to `api` module
