+++ exonum/src/api/public/mod.rs
@@ -15,7 +15,7 @@
 //! Public part of the Exonum rest api.
 
 pub use self::system::{HealthCheckInfo, SystemApi};
-pub use self::blockchain_explorer::ExplorerApi;
+pub use self::blockchain_explorer::{BlocksRange, ExplorerApi};
 
 mod system;
 mod blockchain_explorer;
