+++ testkit/src/lib.rs
@@ -23,10 +23,10 @@
 //! extern crate exonum_testkit;
 //! extern crate serde_json;
 //!
+//! use exonum::api::public::BlocksRange;
 //! use exonum::blockchain::{Block, Schema, Service, Transaction, TransactionSet, ExecutionResult};
 //! use exonum::crypto::{gen_keypair, Hash, PublicKey, CryptoHash};
 //! use exonum::encoding;
-//! use exonum::explorer::BlocksRange;
 //! use exonum::helpers::Height;
 //! use exonum::messages::{Message, RawTransaction};
 //! use exonum::storage::{Snapshot, Fork};
